# openapi.yml
openapi: 3.0.3
info:
  title: Corazón Migrante API
  version: 1.0.0
  description: "API pública y administrativa de Corazón Migrante.\n\nPrefijo global: `/api/v1`.\n\nContrato de respuesta:\n\
    - Respuesta simple: `{ data, meta }`.\n- Respuesta paginada: `{ data, pagination, meta }`.\n- Error: `{ error: { code,\
    \ message, details }, meta }`.\n\nSeguridad:\n- Los endpoints públicos están marcados como `public` en su descripción.\n\
    - Los endpoints privados usan `Bearer JWT`.\n- Algunos endpoints administrativos requieren permisos específicos además\
    \ del rol.\n\nDisponibilidad de terapeutas:\n- `GET /booking/availability` calcula los horarios disponibles por terapeuta.\n\
    - Parte de los horarios recurrentes configurados en `therapist_schedules`.\n- Resta bloqueos activos de agenda (`therapist_blocked_times`).\n\
    - Resta citas activas/ocupadas (`appointments` con status `REQUESTED` o `CONFIRMED`).\n- El rango máximo soportado es\
    \ 31 días.\n\nLímites de uso (rate limiting):\n- Límite global por IP: 120 solicitudes / 60s. Al excederlo, la API responde\
    \ `429`\n  con `{ error: { code: \"HTTP_429\", ... } }`.\n- `POST /auth/login`, `/auth/register/*`: 10 solicitudes / 60s\
    \ por IP.\n- `POST /auth/password-reset/request` y `/confirm`: 5 solicitudes / 60s por IP.\n- El cuerpo de las solicitudes\
    \ JSON está limitado a 2 MB (`413` si se excede).\n"
servers:
- url: https://api.corazondemigrante.com/api/v1
  description: Producción
...
security:
- bearerAuth: []
paths:
...
  /me:
    get:
      tags:
      - Users
      summary: Obtener perfil de sesión
      description: Requiere autenticación. Devuelve usuario, rol, permisos y perfiles asociados.
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /me/patient-profile:
    patch:
      tags:
      - Users
      summary: Actualizar perfil de paciente
      description: Requiere rol `PATIENT`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientProfileRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /me/therapist-profile:
    patch:
      tags:
      - Users
      summary: Actualizar perfil de terapeuta
      description: Requiere rol `THERAPIST`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTherapistProfileRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /admin/users:
    get:
      tags:
      - Users
      summary: Listar usuarios
      description: Requiere rol `ADMIN` o `SUPER_ADMIN` y permiso `users:read`.
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PPage'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/PLimit'
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/Search'
      - $ref: '#/components/parameters/Q'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/SortBy'
      - $ref: '#/components/parameters/Order'
      - $ref: '#/components/parameters/SortDir'
      - $ref: '#/components/parameters/Status'
      - $ref: '#/components/parameters/Estado'
      - $ref: '#/components/parameters/PEstado'
      - $ref: '#/components/parameters/Role'
      - $ref: '#/components/parameters/Rol'
      - $ref: '#/components/parameters/PRol'
      - $ref: '#/components/parameters/TipoUsuario'
      responses:
        '200':
          $ref: '#/components/responses/PaginatedEnvelope'
...
  /therapists/me/schedules:
    get:
      tags:
      - Scheduling
      summary: Listar horarios recurrentes del terapeuta autenticado
      description: Requiere rol `THERAPIST`. Devuelve los bloques base de atención configurados por día de semana.
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags:
      - Scheduling
      summary: Crear horario recurrente del terapeuta autenticado
      description: Requiere rol `THERAPIST`. Rechaza solapamientos contra horarios activos existentes del mismo terapeuta.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
            example:
              weekday: 1
              startTime: 09:00
              endTime: '13:00'
              timezone: America/La_Paz
              effectiveFrom: '2026-07-01'
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /therapists/me/blocked-times:
    post:
      tags:
      - Scheduling
      summary: Crear bloqueo de agenda del terapeuta autenticado
      description: 'Requiere rol `THERAPIST`.

        Estos bloqueos se descuentan del cálculo público de disponibilidad en `/booking/availability`.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlockedTimeRequest'
            example:
              startAt: '2026-07-10T13:00:00.000-04:00'
              endAt: '2026-07-10T15:00:00.000-04:00'
              reason: Supervisión clínica
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
  /booking/availability:
    get:
      tags:
      - Booking
      summary: Obtener horarios disponibles por terapeuta y producto
      description: 'Público. Calcula slots disponibles para reservar una cita.


        Algoritmo funcional confirmado:

        1. Busca horarios recurrentes activos del terapeuta en `therapist_schedules`.

        2. Usa la duración del producto terapéutico (`durationMinutes`) para partir el horario en slots.

        3. Descarta slots que se solapan con bloqueos activos de agenda (`therapist_blocked_times`).

        4. Descarta slots que se solapan con citas activas (`appointments` con status `REQUESTED` o `CONFIRMED`).

        5. Cuando existe la vista `v_therapist_unavailable_intervals`, se usa como fuente consolidada de intervalos no disponibles.

        6. Devuelve solo slots libres entre `from` y `to`.


        El rango máximo permitido es de 31 días.

        '
      security: []
      parameters:
      - name: therapistUserId
        in: query
        required: true
        schema:
          type: string
          format: uuid
        description: ID del usuario terapeuta.
      - name: productId
        in: query
        required: true
        schema:
          type: string
          format: uuid
        description: ID del producto terapéutico. Define duración del slot.
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date
          example: '2026-07-01'
        description: Fecha inicial local incluida.
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date
          example: '2026-07-14'
        description: Fecha final local incluida.
      - name: timezone
        in: query
        required: true
        schema:
          type: string
          example: America/La_Paz
        description: Zona horaria para interpretar horarios recurrentes.
      responses:
        '200':
          description: Slots disponibles calculados.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/EnvelopeObject'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/AvailabilityResponse'
              example:
                data:
                  therapistUserId: 7d2f75d0-5298-4a92-87e0-3aa7a4371455
                  productId: 7409c34e-8470-42df-8eb2-83f579d87432
                  slots:
                  - startAt: '2026-07-06T13:00:00.000Z'
                    endAt: '2026-07-06T14:00:00.000Z'
                    timezone: America/La_Paz
                meta:
                  requestId: 2abf44ee-2b5a-4a49-a723-f8bb1639b7bd
                  timestamp: '2026-07-05T15:13:13.536Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Página actual. Tiene prioridad sobre `p_page`.
    PPage:
      name: p_page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
      description: Alias legacy aceptado para paginación. `page` tiene prioridad cuando ambos llegan.
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Tamaño de página histórico.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Tamaño de página recomendado para frontend/smoke. Tiene prioridad sobre `pageSize` y `p_limit`.
    PLimit:
      name: p_limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Alias legacy aceptado para tamaño de página. `limit` y `pageSize` tienen prioridad cuando llegan juntos.
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
      description: Búsqueda textual canónica.
    Q:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: Alias corto para búsqueda. `search` tiene prioridad cuando ambos llegan.
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        default: createdAt
      description: Campo de orden. El backend normaliza alias camelCase/snake_case seguros y evita ORDER BY con columnas inexistentes.
    SortBy:
      name: sortBy
      in: query
      required: false
      schema:
        type: string
      description: Alias frontend para `sort`. `sort` tiene prioridad si ambos llegan.
    Order:
      name: order
      in: query
      required: false
      schema:
        type: string
        enum:
        - asc
        - desc
        default: desc
    SortDir:
      name: sortDir
      in: query
      required: false
      schema:
        type: string
        enum:
        - asc
        - desc
        default: desc
      description: Alias frontend para `order`. `order` tiene prioridad si ambos llegan.
    Status:
      name: status
      in: query
      required: false
      schema:
        type: string
      description: Filtro canónico de estado. En usuarios acepta ACTIVE/INACTIVE/BLOCKED/PENDING y aliases activo/inactivo/bloqueado/pendiente.
    Estado:
      name: estado
      in: query
      required: false
      schema:
        type: string
      description: Alias español para `status`.
    PEstado:
      name: p_estado
      in: query
      required: false
      schema:
        type: string
      description: Alias legacy español para `status`.
    Role:
      name: role
      in: query
      required: false
      schema:
        type: string
      description: Filtro canónico de rol. En usuarios acepta THERAPIST/TERAPEUTA, PATIENT/PACIENTE, ACCOUNTANT/CONTADOR,
        ADMIN y SUPER_ADMIN.
    Rol:
      name: rol
      in: query
      required: false
      schema:
        type: string
      description: Alias español para `role`.
    PRol:
      name: p_rol
      in: query
      required: false
      schema:
        type: string
      description: Alias legacy español para `role`.
    TipoUsuario:
      name: tipo_usuario
      in: query
      required: false
      schema:
        type: string
      description: Alias legacy para `role`.
  responses:
    OkEnvelope:
      description: Respuesta exitosa envuelta en `{ data, meta }`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeObject'
    CreatedEnvelope:
      description: Recurso creado. Respuesta envuelta en `{ data, meta }`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeObject'
    PaginatedEnvelope:
      description: Lista paginada envuelta en `{ data, pagination, meta }`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PaginatedEnvelope'
    ValidationError:
      description: Error de validación o contrato de entrada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: HTTP_400
              message:
              - property debug should not exist
              details:
              - property debug should not exist
            meta:
              requestId: 315c2096-df71-447a-92ad-2f9329316896
              timestamp: '2026-07-05T15:13:13.536Z'
    UnauthorizedError:
      description: Token ausente, vencido o inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ForbiddenError:
      description: Usuario autenticado sin rol o permiso suficiente.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFoundError:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Meta:
      type: object
      required:
      - requestId
      - timestamp
      properties:
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    EnvelopeObject:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          oneOf:
          - type: object
            nullable: true
          - type: array
            items: {}
          - type: string
          - type: boolean
          - type: number
        meta:
          $ref: '#/components/schemas/Meta'
    Pagination:
      type: object
      required:
      - page
      - pageSize
      - limit
      - total
      - totalPages
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 138
        totalPages:
          type: integer
          example: 7
    PaginatedEnvelope:
      type: object
      required:
      - data
      - pagination
      - meta
      properties:
        data:
          type: array
          items: {}
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorEnvelope:
      type: object
      required:
      - error
      - meta
      properties:
        error:
          type: object
          required:
          - code
          - message
          - details
          properties:
            code:
              type: string
              example: HTTP_400
            message:
              oneOf:
              - type: string
              - type: array
                items:
                  type: string
            details:
              type: array
              items: {}
        meta:
          $ref: '#/components/schemas/Meta'
...
    UpdatePatientProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        birthDate:
          type: string
          format: date
        country:
          type: string
        city:
          type: string
        occupation:
          type: string
    UpdateTherapistProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        title:
          type: string
        mainSpecialty:
          type: string
        bio:
          type: string
        personalPhrase:
          type: string
...
    CreateScheduleRequest:
      type: object
      required:
      - weekday
      - startTime
      - endTime
      - timezone
      - effectiveFrom
      properties:
        weekday:
          type: integer
          minimum: 0
          maximum: 6
          description: Día de semana. 0 domingo, 1 lunes, ..., 6 sábado.
        startTime:
          type: string
          pattern: ^\d{2}:\d{2}$
          example: 09:00
        endTime:
          type: string
          pattern: ^\d{2}:\d{2}$
          example: '13:00'
        timezone:
          type: string
          example: America/La_Paz
        effectiveFrom:
          type: string
          format: date
          example: '2026-07-01'
        effectiveTo:
          type: string
          format: date
          nullable: true
    CreateBlockedTimeRequest:
      type: object
      required:
      - startAt
      - endAt
      properties:
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        reason:
          type: string
    AvailabilityResponse:
      type: object
      required:
      - therapistUserId
      - productId
      - slots
      properties:
        therapistUserId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        slots:
          type: array
          items:
            type: object
            required:
            - startAt
            - endAt
            - timezone
            properties:
              startAt:
                type: string
                format: date-time
              endAt:
                type: string
                format: date-time
              timezone:
                type: string
                example: America/La_Paz
...